feat: move to usercontroller
